requirements.txt
faiss-cpu==1.13.2
sentence-transformers==5.3.0
rank-bm25==0.2.2
langchain-text-splitters==1.1.1
ragas==0.4.3
tiktoken==0.12.0
numpy==1.26.4
pandas==2.1.4
seaborn==0.13.2
matplotlib==3.8.2
tqdm==4.67.3
PyYAML==6.0.1
ollama==0.6.1
requests==2.32.5
